+++ src/weather/roll-weather.ts
@@ -28,7 +28,7 @@ async function rollOnWeatherEventTable(
     } else {
         await table.toMessage(results, {roll: draw.roll, messageOptions: {rollMode}});
         if (rollTwice) {
-            await postMessage('Choose a second weather event!');
+            await postMessage('Choose a second weather event!', rollMode !== 'publicroll');
         }
     }
 }
@@ -58,8 +58,8 @@ async function rollCheck(dc: number, flavor: string, rollMode: RollMode): Promis
     return {isSuccess, total: roll.total};
 }
 
-async function postMessage(message: string): Promise<void> {
-    await ChatMessage.create({content: message, blind: true});
+async function postMessage(message: string, blind: boolean): Promise<void> {
+    await ChatMessage.create({content: message, blind});
 }
 
 export async function rollKingmakerWeather(game: Game): Promise<void> {
@@ -109,5 +109,5 @@ async function rollWeather(game: Game, averagePartyLevel: number, weatherHazardR
         }
     }
     await rollWeatherEvent(game, averagePartyLevel, weatherHazardRange, rollMode);
-    await postMessage(message);
+    await postMessage(message, rollMode !== 'publicroll');
 }